retrained model on large data

+++ backend/routes/prediction.py
@@ -10,13 +10,13 @@ from config import UPLOAD_FOLDER
 from utils.validators import allowed_file
 from services.prediction_service import predict_image
 
+
 prediction_bp = Blueprint("prediction", __name__)
 
 
 @prediction_bp.route("/predict", methods=["POST"])
 def predict():
 
-    # Check file exists
     if "image" not in request.files:
         return jsonify({
             "error": "No image uploaded."
@@ -24,19 +24,16 @@ def predict():
 
     file = request.files["image"]
 
-    # Empty filename
     if file.filename == "":
         return jsonify({
             "error": "Please choose an image."
         }), 400
 
-    # Invalid extension
     if not allowed_file(file.filename):
         return jsonify({
-            "error": "Only JPG, JPEG and PNG are allowed."
+            "error": "Only JPG, JPEG and PNG images are allowed."
         }), 400
 
-    # Secure filename
     filename = secure_filename(file.filename)
 
     image_path = os.path.join(
@@ -44,13 +41,11 @@ def predict():
         filename
     )
 
-    # Save image
     file.save(image_path)
 
-    # Predict
     result = predict_image(image_path)
 
-    # Delete uploaded file
-    os.remove(image_path)
+    if os.path.exists(image_path):
+        os.remove(image_path)
 
     return jsonify(result)
